Removed interface for customer
