+++ src/main/java/library/customer/CustomerBuilder.java
@@ -2,33 +2,33 @@ package library.customer;
 
 import java.time.LocalDate;
 
-public class LibraryCustomerBuilder {
+public class CustomerBuilder {
 
     private String[] name;
     private String email;
     private LocalDate dateOfBirth;
     private Address address;
 
-    public LibraryCustomer buildCustomer() {
-        return new LibraryCustomer(name, email, dateOfBirth, address);
+    public Customer buildCustomer() {
+        return new Customer(name, email, dateOfBirth, address);
     }
 
-    public LibraryCustomerBuilder name(String forename, String surname) {
+    public CustomerBuilder name(String forename, String surname) {
         this.name = new String[]{forename, surname};
         return this;
     }
 
-    public LibraryCustomerBuilder email(String email) {
+    public CustomerBuilder email(String email) {
         this.email = email;
         return this;
     }
 
-    public LibraryCustomerBuilder dob(LocalDate dateOfBirth) {
+    public CustomerBuilder dob(LocalDate dateOfBirth) {
         this.dateOfBirth = dateOfBirth;
         return this;
     }
 
-    public LibraryCustomerBuilder address(Address address) {
+    public CustomerBuilder address(Address address) {
         this.address = new Address(address.getStreetAddress(), address.getPostCode());
         if (address.getTown() != null) {
             this.address.setTown(address.getTown());
